+++ include/vb/impl/Constellation0.hxx
@@ -327,13 +327,13 @@ namespace vb {
 
         T rad = mindist / 5;
 
-        std::vector<std::vector<cplx>>     hands;
-        std::vector<std::vector<unsigned>> halfedges;
-        unsigned                           index = 0;
+        std::vector<std::vector<cplx>>   hands;
+        std::vector<std::vector<size_t>> halfedges;
+        unsigned                         index = 0;
 
         for (Star<T> z : Z) {
-            std::vector<cplx>     hs;
-            std::vector<unsigned> he;
+            std::vector<cplx>   hs;
+            std::vector<size_t> he;
 
             cplx u = z.z + rad * exp(cplx(0, .001));
             T    s = imag((*this)(u));
@@ -351,10 +351,10 @@ namespace vb {
             halfedges.push_back(he);
         }
 
-        std::vector<cplx>     hs;
-        std::vector<unsigned> he;
-        cplx                  u = large * exp(cplx(0, .001));
-        T                     s = imag((*this)(u));
+        std::vector<cplx>   hs;
+        std::vector<size_t> he;
+        cplx                u = large * exp(cplx(0, .001));
+        T                   s = imag((*this)(u));
         for (unsigned i = 0; i < 10 * maxdeg; ++i) {
             u    = exp(cplx(0, -2 * pi_<T>() / (10 * maxdeg))) * u;
             T ns = imag((*this)(u));
@@ -367,7 +367,7 @@ namespace vb {
         hands.push_back(hs);
         halfedges.push_back(he);
 
-        std::vector<std::vector<unsigned>> pairs;
+        std::vector<std::vector<size_t>> pairs;
 
         for (unsigned i = 0; i < Z.size(); ++i) {
             for (unsigned j = 0; j < hands[i].size(); ++j) {
@@ -389,7 +389,7 @@ namespace vb {
                             }
                         }
                         if (!h) return std::nullopt;
-                        pairs.emplace_back(std::vector<unsigned>{halfedges[i][j], halfedges[k][*h]});
+                        pairs.emplace_back(std::vector<size_t>{halfedges[i][j], halfedges[k][*h]});
                         looking = false;
                     }
                     for (unsigned k = 0; k < Z.size(); ++k)
@@ -405,7 +405,7 @@ namespace vb {
                             }
                             if (!h) return std::nullopt;
                             if (halfedges[i][j] < halfedges[k][*h])
-                                pairs.emplace_back(std::vector<unsigned>{halfedges[i][j], halfedges[k][*h]});
+                                pairs.emplace_back(std::vector<size_t>{halfedges[i][j], halfedges[k][*h]});
                             looking = false;
                         }
                     if (sl * imag((*this)(nz)) > 0)
